backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
pypdf==4.3.1
python-docx==1.1.2
requests>=2.31.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
